MFC: fixes for OpenSSL ASN.1 invalid memory accesses (CVE-2009-0590 and CVE-2009-0789).

+++ sys/compat/ibcs2/ibcs2_types.h
@@ -1,3 +1,4 @@
+/*	$OpenBSD: ibcs2_types.h,v 1.2 1996/08/02 20:35:16 niklas Exp $	*/
 /*	$NetBSD: ibcs2_types.h,v 1.5 1995/08/14 01:11:54 mycroft Exp $	*/
 
 /*
@@ -37,7 +38,7 @@ typedef unsigned char	ibcs2_uchar_t;
 typedef unsigned long	ibcs2_ulong_t;
 
 typedef char *		ibcs2_caddr_t;
-typedef long		ibcs2_daddr_t;
+typedef int32_t		ibcs2_daddr_t;
 typedef long		ibcs2_off_t;
 typedef long		ibcs2_key_t;
 typedef unsigned short	ibcs2_uid_t;
